Move Singleton to the other file
